🔄 synced local 'skyvern/' with remote 'skyvern/'
Replacing
```
                scraped_page = await scrape_website(
                    browser_state,
                    url,
                    app.AGENT_FUNCTION.cleanup_element_tree_factory(),
                    scrape_exclude=app.scrape_exclude,
                )
```
with
```
                scraped_page = await browser_state.scrape_website(
                    url=url,
                    cleanup_element_tree=app.AGENT_FUNCTION.cleanup_element_tree_factory(),
                    scrape_exclude=app.scrape_exclude,
                )
```
(`RealBrowserState` delegates the call to the scraper)

This way in tests we will be able to hook up fake scraping logic.

Pure refactoring, no functional changes.

<!-- ELLIPSIS_HIDDEN -->

> [!IMPORTANT]
> Extract `scrape_website` to `BrowserState` and move `ScrapedPage` to `scraped_page.py`, updating imports and method calls across the codebase.
>
>   - **Behavior**:
>     - Extract `scrape_website` method from `scraper.py` to `BrowserState` class in `real_browser_state.py`.
>     - Update calls to `scrape_website` in `script_skyvern_page.py`, `agent.py`, and `task_v2_service.py` to use `browser_state.scrape_website()`.
>   - **Refactoring**:
>     - Move `ScrapedPage` class and related functions to `scraped_page.py`.
>     - Update imports in `actions.py`, `action_click.py`, `action_complete.py`, `action_input.py`, `action_select_option.py`, `action_upload_file.py`, `click.py`, `url.py`, `replay_task_actions.py`, `real_skyvern_page_ai.py`, `script_skyvern_page.py`, `agent.py`, `agent_functions.py`, `script_service.py`, `task_v2_service.py`, `prompt_engine.py`, `caching.py`, `handler.py`, `models.py`, `parse_actions.py`, `browser_state.py`, `real_browser_state.py`, `dom.py`, and `test_parallel_verification.py` to reflect new file structure.
>   - **Misc**:
>     - Rename `skyvern/webeye/scraper/scraper.py` to `skyvern/webeye/scraper/scraped_page.py` for better organization.
>
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for 0e229358a6922f442ceb1c20ae8d3349931f5306. You can [customize](https://app.ellipsis.dev/Skyvern-AI/settings/summaries) this summary. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->
