+++ skyvern/forge/agent_functions.py
@@ -24,7 +24,7 @@ from skyvern.services import workflow_script_service
 from skyvern.webeye.actions.action_types import POST_ACTION_EXECUTION_ACTION_TYPES
 from skyvern.webeye.actions.actions import Action
 from skyvern.webeye.browser_state import BrowserState
-from skyvern.webeye.scraper.scraper import ELEMENT_NODE_ATTRIBUTES, CleanupElementTreeFunc, json_to_html
+from skyvern.webeye.scraper.scraped_page import ELEMENT_NODE_ATTRIBUTES, CleanupElementTreeFunc, json_to_html
 from skyvern.webeye.utils.dom import SkyvernElement
 from skyvern.webeye.utils.page import SkyvernFrame
